+++ trunk/p2p_slotion/asyncrtspclientsocket.h
@@ -50,7 +50,7 @@ class RTSPClientSocket : public ProxySocketBegin
 public:
   explicit RTSPClientSocket(ProxyP2PSession *proxy_p2p_session,
     talk_base::AsyncSocket *int_socket,
-    uint32 server_socket_number,
+    intptr_t server_socket_number,
     const talk_base::SocketAddress &server_addr);
 private:
   void OnInternalConnect(talk_base::AsyncSocket* socket);
@@ -58,7 +58,7 @@ private:
   //virtual void OnInternalClose(talk_base::AsyncSocket* socket,
   //  int err);
 private:
-  uint32                       server_socket_number_;
+  intptr_t                    server_socket_number_;
   DISALLOW_EVIL_CONSTRUCTORS(RTSPClientSocket);
 };
